refactor : merge glsl
