+++ src/crossGrid/CrossGridMaterial.ts
@@ -1,4 +1,4 @@
-import VertexShader from "../ShaderPhongMaterial.vert.glsl";
+import { vertex } from "../ShaderPhongMaterial.glsl";
 import { WavyGridMaterial } from "../WavyGridMaterial";
 import { fragment } from "./CrossGridMaterial.glsl";
 import { ShaderMaterialParameters, UniformsUtils } from "three";
@@ -26,7 +26,7 @@ export class CrossGridMaterial extends WavyGridMaterial {
   }
 
   constructor(parameters?: ShaderMaterialParameters) {
-    super(VertexShader(), fragment, parameters);
+    super(vertex, fragment, parameters);
   }
 
   protected initUniforms(): void {
