+++ src/halftoneGrid/HalftoneGridMaterial.ts
@@ -1,4 +1,4 @@
-import VertexShader from "../ShaderPhongMaterial.vert.glsl";
+import { vertex } from "../ShaderPhongMaterial.glsl";
 import { WavyGridMaterial } from "../WavyGridMaterial";
 import { fragment } from "./HalftoneGridMaterial.glsl";
 import { ShaderMaterialParameters, UniformsUtils } from "three";
@@ -15,7 +15,7 @@ export class HalftoneGridMaterial extends WavyGridMaterial {
   }
 
   constructor(parameters?: ShaderMaterialParameters) {
-    super(VertexShader(), fragment, parameters);
+    super(vertex, fragment, parameters);
   }
 
   protected initUniforms(): void {
